feat(boilerplate): Toggle component refactored to Checkbox, Radio and Switch (#2667 by @frankcalise)
* refactor(toggle): split to Checkbox, Radio and Switch components

* fix(boilerplate): update demos to use specific toggle components

* updating deprecated reanimated Extrapolate

* TS fixes

* docs: explode Toggle docs to Checkbox, Radio, Switch

* docs(components): added Checkbox, Radio, Switch to main listing

* fix(boilerplate): remove reanimated from Checkbox

* fix(boilerplate): remove reanimated from Radio

* fix(boilerplate): checkbox, radio match old duration times from default reanimated values

* fix(boilerplate): remove reanimated dep from Switch

* docs(toggle): tsx formatting

* refactor(boilerplate):drop prefix from component specific props

* some missed props

* fix(boilerplate): toggle components accessibilityRole

---------
