+++ boilerplate/app/components/Toggle/Checkbox.tsx
@@ -1,0 +1,118 @@
+import React, { useState, useEffect } from "react"
+import { Image, ImageStyle, Animated, StyleProp, View, ViewStyle } from "react-native"
+import { colors } from "../../theme"
+import { iconRegistry, IconTypes } from "../Icon"
+import { $inputOuterBase, BaseToggleInputProps, ToggleProps, Toggle } from "./Toggle"
+
+export interface CheckboxToggleProps extends Omit<ToggleProps<CheckboxInputProps>, "ToggleInput"> {
+  /**
+   * Optional style prop that affects the Image component.
+   */
+  inputDetailStyle?: ImageStyle
+  /**
+   * Checkbox-only prop that changes the icon used for the "on" state.
+   */
+  icon?: IconTypes
+}
+
+interface CheckboxInputProps extends BaseToggleInputProps<CheckboxToggleProps> {
+  icon?: CheckboxToggleProps["icon"]
+}
+/**
+ * @param {CheckboxToggleProps} props - The props for the `Checkbox` component.
+ * @returns {JSX.Element} The rendered `Checkbox` component.
+ */
+export function Checkbox(props: CheckboxToggleProps) {
+  return <Toggle accessibilityRole="checkbox" {...props} ToggleInput={CheckboxInput} />
+}
+
+function CheckboxInput(props: CheckboxInputProps) {
+  const {
+    on,
+    status,
+    disabled,
+    icon = "check",
+    outerStyle: $outerStyleOverride,
+    innerStyle: $innerStyleOverride,
+    detailStyle: $detailStyleOverride,
+  } = props
+
+  const [opacity] = useState(new Animated.Value(0))
+
+  useEffect(() => {
+    Animated.timing(opacity, {
+      toValue: on ? 1 : 0,
+      duration: 300,
+      useNativeDriver: true,
+    }).start()
+  }, [on])
+
+  const offBackgroundColor = [
+    disabled && colors.palette.neutral400,
+    status === "error" && colors.errorBackground,
+    colors.palette.neutral200,
+  ].filter(Boolean)[0]
+
+  const outerBorderColor = [
+    disabled && colors.palette.neutral400,
+    status === "error" && colors.error,
+    !on && colors.palette.neutral800,
+    colors.palette.secondary500,
+  ].filter(Boolean)[0]
+
+  const onBackgroundColor = [
+    disabled && colors.transparent,
+    status === "error" && colors.errorBackground,
+    colors.palette.secondary500,
+  ].filter(Boolean)[0]
+
+  const iconTintColor = [
+    disabled && colors.palette.neutral600,
+    status === "error" && colors.error,
+    colors.palette.accent100,
+  ].filter(Boolean)[0]
+
+  return (
+    <View
+      style={[
+        $inputOuter,
+        { backgroundColor: offBackgroundColor, borderColor: outerBorderColor },
+        $outerStyleOverride,
+      ]}
+    >
+      <Animated.View
+        style={[
+          $checkboxInner,
+          { backgroundColor: onBackgroundColor },
+          $innerStyleOverride,
+          { opacity },
+        ]}
+      >
+        <Image
+          source={icon ? iconRegistry[icon] : iconRegistry.check}
+          style={[
+            $checkboxDetail,
+            !!iconTintColor && { tintColor: iconTintColor },
+            $detailStyleOverride,
+          ]}
+        />
+      </Animated.View>
+    </View>
+  )
+}
+
+const $checkboxInner: ViewStyle = {
+  width: "100%",
+  height: "100%",
+  alignItems: "center",
+  justifyContent: "center",
+  overflow: "hidden",
+}
+
+const $checkboxDetail: ImageStyle = {
+  width: 20,
+  height: 20,
+  resizeMode: "contain",
+}
+
+const $inputOuter: StyleProp<ViewStyle> = [$inputOuterBase, { borderRadius: 4 }]
