+++ boilerplate/app/components/Toggle/index.ts
@@ -1,0 +1,3 @@
+export * from "./Checkbox"
+export * from "./Radio"
+export * from "./Switch"
