+++ boilerplate/app/screens/DemoPodcastListScreen.tsx
@@ -14,7 +14,7 @@ import {
 } from "react-native"
 import { type ContentStyle } from "@shopify/flash-list"
 import Animated, {
-  Extrapolate,
+  Extrapolation,
   interpolate,
   useAnimatedStyle,
   useSharedValue,
@@ -28,8 +28,8 @@ import {
   Icon,
   ListView,
   Screen,
+  Switch,
   Text,
-  Toggle,
 } from "../components"
 import { isRTL, translate } from "../i18n"
 import { useStores } from "../models"
@@ -111,12 +111,11 @@ export const DemoPodcastListScreen: FC<DemoTabScreenProps<"DemoPodcastList">> = 
               <Text preset="heading" tx="demoPodcastListScreen.title" />
               {(episodeStore.favoritesOnly || episodeStore.episodesForList.length > 0) && (
                 <View style={$toggle}>
-                  <Toggle
+                  <Switch
                     value={episodeStore.favoritesOnly}
                     onValueChange={() =>
                       episodeStore.setProp("favoritesOnly", !episodeStore.favoritesOnly)
                     }
-                    variant="switch"
                     labelTx="demoPodcastListScreen.onlyFavorites"
                     labelPosition="left"
                     labelStyle={$labelStyle}
@@ -159,10 +158,10 @@ const EpisodeCard = observer(function EpisodeCard({
     return {
       transform: [
         {
-          scale: interpolate(liked.value, [0, 1], [1, 0], Extrapolate.EXTEND),
+          scale: interpolate(liked.value, [0, 1], [1, 0], Extrapolation.EXTEND),
         },
       ],
-      opacity: interpolate(liked.value, [0, 1], [1, 0], Extrapolate.CLAMP),
+      opacity: interpolate(liked.value, [0, 1], [1, 0], Extrapolation.CLAMP),
     }
   })
