+++ docs/_category_.json
@@ -1,8 +1,8 @@
 {
-  "label": "Ignite",
+  "label": "ignite-cli",
   "link": null,
   "customProps": {
-    "description": "The battle-tested React Native boilerplate",
+    "description": "",
     "projectName": "ignite-cli",
     "sidebar": {
       "type": "autogenerated",
