+++ docs/boilerplate/app/components/Checkbox.md
@@ -1,0 +1,21 @@
+---
+sidebar_position: 32
+---
+
+import ToggleProps from './\_toggle_props.mdx';
+
+# Checkbox
+
+The `Checkbox` component provides a simple way to collect user input for a boolean value.
+
+## Checkbox Props
+
+### `icon`
+
+The `icon` is a prop for the checkbox variant that allows you to customize the icon used for the "on" state.
+
+```tsx
+<Checkbox icon="ladybug" />
+```
+
+<ToggleProps componentName="Checkbox" />
