+++ docs/boilerplate/app/components/Radio.md
@@ -1,0 +1,11 @@
+---
+sidebar_position: 37
+---
+
+import ToggleProps from './\_toggle_props.mdx';
+
+# Radio
+
+The `Radio` component provides a simple way to collect user input for a boolean value.
+
+<ToggleProps componentName="Radio" />
